requirements.txt
flask==2.1.0
pagarme-python==0.3.3
requests==2.26.0
python-dotenv==0.19.2